add Factories and refactor test setup to conftest file
